--- babyOne/app/src/main/java/com/example/babyone/GraphActivity.java
@@ -1,53 +1,0 @@
-package com.example.babyone;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.graphics.Color;
-import android.os.Bundle;
-
-import com.androidplot.xy.LineAndPointFormatter;
-import com.androidplot.xy.SimpleXYSeries;
-import com.androidplot.xy.XYGraphWidget;
-import com.androidplot.xy.XYPlot;
-import com.androidplot.xy.XYSeries;
-
-import java.text.FieldPosition;
-import java.text.Format;
-import java.text.ParsePosition;
-import java.util.Arrays;
-
-public class GraphActivity extends AppCompatActivity {
-
-    XYPlot plot;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_graph);
-
-        plot = findViewById(R.id.plot);
-
-        final Number[] domainLabels = {1,2,3,4,5,6,7,8,9,10,11,12,13,14,15};
-        Number[] series1Numbers = {1,3,6,15,3,20,34,25,21,38,19,88,75,60,12};
-
-        XYSeries series1 = new SimpleXYSeries(Arrays.asList(series1Numbers),
-                SimpleXYSeries.ArrayFormat.Y_VALS_ONLY,"Series 1");
-
-        LineAndPointFormatter series1Format = new LineAndPointFormatter(Color.RED,Color.GREEN,null,null);
-
-        plot.addSeries(series1,series1Format);
-
-        plot.getGraph().getLineLabelStyle(XYGraphWidget.Edge.BOTTOM).setFormat(new Format() {
-            @Override
-            public StringBuffer format(Object obj, StringBuffer toAppendTo, FieldPosition pos) {
-                int i = Math.round( ((Number)obj).floatValue() );
-                return toAppendTo.append(domainLabels[i]);
-            }
-
-            @Override
-            public Object parseObject(String source, ParsePosition pos) {
-                return null;
-            }
-        });
-    }
-}
