+++ babyOne/app/src/main/java/com/example/babyone/profileFragment.java
@@ -1,12 +1,27 @@
 package com.example.babyone;
 
+import android.graphics.Color;
+import android.graphics.Paint;
 import android.os.Bundle;
 
+import androidx.cardview.widget.CardView;
 import androidx.fragment.app.Fragment;
 
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.androidplot.xy.LineAndPointFormatter;
+import com.androidplot.xy.SimpleXYSeries;
+import com.androidplot.xy.XYGraphWidget;
+import com.androidplot.xy.XYPlot;
+import com.androidplot.xy.XYSeries;
+import com.google.android.material.card.MaterialCardView;
+
+import java.text.FieldPosition;
+import java.text.Format;
+import java.text.ParsePosition;
+import java.util.Arrays;
 
 /**
  * A simple {@link Fragment} subclass.
@@ -15,6 +30,7 @@ import android.view.ViewGroup;
  */
 public class profileFragment extends Fragment {
 
+    private MaterialCardView cardView_2;
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -53,12 +69,58 @@ public class profileFragment extends Fragment {
             mParam1 = getArguments().getString(ARG_PARAM1);
             mParam2 = getArguments().getString(ARG_PARAM2);
         }
+
     }
+
+    XYPlot plot;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_profile, container, false);
+
+        View view = inflater.inflate(R.layout.fragment_profile, container, false);
+
+        plot = view.findViewById(R.id.plot);
+
+        plot.setDomainTitle(null);
+        plot.setRangeTitle(null);
+
+        plot.getGraph().getBackgroundPaint().setColor(Color.parseColor("#9DCBE1"));
+
+        String[] domainLabels = {"0","mon","feb","mar","apr","may","jun","jul","aug","sep","oct","nov","dec",};
+        Number[] series1Numbers = {0,2,3,4,2,3,5,4,6,3,4,2,1};
+
+        XYSeries series1 = new SimpleXYSeries(Arrays.asList(series1Numbers),
+                SimpleXYSeries.ArrayFormat.Y_VALS_ONLY,null);
+
+        LineAndPointFormatter series1Format = new LineAndPointFormatter(Color.rgb(255, 165, 0),null,null,null);
+
+        series1Format.setFillPaint(new Paint(Paint.ANTI_ALIAS_FLAG));
+        series1Format.getFillPaint().setStyle(Paint.Style.FILL);
+        series1Format.getFillPaint().setColor(Color.rgb(255, 165, 0)); // Set fill color to orange
+
+        plot.addSeries(series1,series1Format);
+
+        plot.getGraph().getLineLabelStyle(XYGraphWidget.Edge.BOTTOM).setFormat(new Format() {
+            @Override
+            public StringBuffer format(Object obj, StringBuffer toAppendTo, FieldPosition pos) {
+                int i = Math.round( ((Number)obj).floatValue() );
+                return toAppendTo.append(domainLabels[i]);
+            }
+
+            @Override
+            public Object parseObject(String source, ParsePosition pos) {
+                return null;
+            }
+        });
+
+        plot.getGraph().getGridBackgroundPaint().setColor(Color.TRANSPARENT);
+        plot.getGraph().getDomainGridLinePaint().setColor(Color.TRANSPARENT);
+        plot.getGraph().getRangeGridLinePaint().setColor(Color.TRANSPARENT);
+        plot.getGraph().getDomainOriginLinePaint().setColor(Color.TRANSPARENT);
+        plot.getGraph().getRangeOriginLinePaint().setColor(Color.TRANSPARENT);
+
+
+        return view;
     }
 }
