--- babyOne/app/src/main/res/layout/activity_graph.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:ap="http://schemas.android.com/apk/res-auto"
-    tools:context=".GraphActivity">
-
-    <com.androidplot.xy.XYPlot
-        android:id="@+id/plot"
-        style="@style/APDefacto.Light"
-        android:layout_width="fill_parent"
-        android:layout_height="fill_parent"
-        ap:domainTitle="domain"
-        ap:rangeTitle="range"
-        ap:title="simple XY Plot"
-        />
-
-</LinearLayout>
